requirements.txt
addict==2.4.0
aider-chat==0.54.4
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==2.0.1
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudscraper==1.2.71
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.1
crcmod==1.7
cryptography==42.0.7
cssselect2==0.7.0
curl_cffi==0.7.3
cycler==0.12.1
datasets==2.18.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
executing==2.0.1
fastapi==0.115.5
fastapi-cache2==0.2.2
filelock==3.15.4
flake8==7.1.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
g4f==0.3.3.3
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grep-ast==0.3.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
imageio==2.34.1
importlib_metadata==7.2.1
importlib_resources==6.4.4
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
kombu==5.4.2
lazy_loader==0.4
litellm==1.44.7
loguru==0.7.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mkl==2021.4.0
mmcv-full==1.7.2
mmdet==3.3.0
mmengine==0.10.4
model-index==0.1.11
modelscope==1.14.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mutagen==1.47.0
mysqlclient==2.2.6
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
onnxruntime==1.18.0
openai==1.42.0
opencv-python==4.10.0.82
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.21.0
prompt_toolkit==3.0.47
protobuf==5.27.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
pygtrans==1.6.1
PyMySQL==1.1.1
pypandoc==1.13
pyparsing==3.1.2
pyperclip==1.9.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2023.4
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
redis==4.6.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.41.2
sympy==1.12.1
tabulate==0.9.0
tbb==2021.12.0
termcolor==2.4.0
terminaltables==3.1.10
tifffile==2024.5.22
tiktoken==0.7.0
time-machine==2.16.0
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.22.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
win32-setctime==1.1.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
yt-dlp==2024.7.2
zipp==3.20.1
